+++ src/main/java/com/vendhq/api/TaxesApi.java
@@ -7,7 +7,6 @@ import com.vendhq.api.Pair;
 
 import javax.ws.rs.core.GenericType;
 
-import java.math.BigDecimal;
 import com.vendhq.models.TaxCollection;
 
 import java.util.ArrayList;
@@ -44,7 +43,7 @@ public class TaxesApi {
    * @return TaxCollection
    * @throws ApiException if fails to make API call
    */
-  public TaxCollection listTaxes(BigDecimal after, BigDecimal before, BigDecimal pageSize) throws ApiException {
+  public TaxCollection listTaxes(Long after, Long before, Integer pageSize) throws ApiException {
     Object localVarPostBody = null;
     
     // create path and map variables
