+++ src/main/java/com/vendhq/models/BrandSample.java
@@ -14,6 +14,7 @@
 package com.vendhq.models;
 
 import java.util.Objects;
+import java.util.Arrays;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
@@ -35,7 +36,7 @@ public class BrandSample {
   private String deletedAt = null;
 
   @JsonProperty("version")
-  private Integer version = null;
+  private Long version = null;
 
   public BrandSample id(String id) {
     this.id = id;
@@ -91,7 +92,7 @@ public class BrandSample {
     this.deletedAt = deletedAt;
   }
 
-  public BrandSample version(Integer version) {
+  public BrandSample version(Long version) {
     this.version = version;
     return this;
   }
@@ -101,11 +102,11 @@ public class BrandSample {
    * @return version
   **/
   @ApiModelProperty(value = "Auto-incrementing object version number.")
-  public Integer getVersion() {
+  public Long getVersion() {
     return version;
   }
 
-  public void setVersion(Integer version) {
+  public void setVersion(Long version) {
     this.version = version;
   }
 
@@ -154,6 +155,6 @@ public class BrandSample {
     }
     return o.toString().replace("\n", "\n    ");
   }
-  
+
 }
