+++ src/main/java/com/vendhq/models/Consignment.java
@@ -14,6 +14,7 @@
 package com.vendhq.models;
 
 import java.util.Objects;
+import java.util.Arrays;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
@@ -80,7 +81,7 @@ public class Consignment {
   private String deletedAt = null;
 
   @JsonProperty("version")
-  private Integer version = null;
+  private Long version = null;
 
   @JsonProperty("total_count_gain")
   private Double totalCountGain = null;
@@ -409,7 +410,7 @@ public class Consignment {
     this.deletedAt = deletedAt;
   }
 
-  public Consignment version(Integer version) {
+  public Consignment version(Long version) {
     this.version = version;
     return this;
   }
@@ -419,11 +420,11 @@ public class Consignment {
    * @return version
   **/
   @ApiModelProperty(value = "Auto-incrementing object version number.")
-  public Integer getVersion() {
+  public Long getVersion() {
     return version;
   }
 
-  public void setVersion(Integer version) {
+  public void setVersion(Long version) {
     this.version = version;
   }
 
@@ -522,6 +523,6 @@ public class Consignment {
     }
     return o.toString().replace("\n", "\n    ");
   }
-  
+
 }
