+++ src/main/java/com/vendhq/models/CustomerGroup.java
@@ -14,6 +14,7 @@
 package com.vendhq.models;
 
 import java.util.Objects;
+import java.util.Arrays;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
@@ -44,7 +45,7 @@ public class CustomerGroup {
   private String deletedAt = null;
 
   @JsonProperty("version")
-  private Integer version = null;
+  private Long version = null;
 
   public CustomerGroup name(String name) {
     this.name = name;
@@ -154,7 +155,7 @@ public class CustomerGroup {
     this.deletedAt = deletedAt;
   }
 
-  public CustomerGroup version(Integer version) {
+  public CustomerGroup version(Long version) {
     this.version = version;
     return this;
   }
@@ -164,11 +165,11 @@ public class CustomerGroup {
    * @return version
   **/
   @ApiModelProperty(value = "Auto-incrementing object version number.")
-  public Integer getVersion() {
+  public Long getVersion() {
     return version;
   }
 
-  public void setVersion(Integer version) {
+  public void setVersion(Long version) {
     this.version = version;
   }
 
@@ -223,6 +224,6 @@ public class CustomerGroup {
     }
     return o.toString().replace("\n", "\n    ");
   }
-  
+
 }
