+++ src/main/java/com/vendhq/models/ImageSample.java
@@ -14,6 +14,7 @@
 package com.vendhq.models;
 
 import java.util.Objects;
+import java.util.Arrays;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
@@ -32,7 +33,7 @@ public class ImageSample {
   private String url = null;
 
   @JsonProperty("version")
-  private Integer version = null;
+  private Long version = null;
 
   public ImageSample id(String id) {
     this.id = id;
@@ -70,7 +71,7 @@ public class ImageSample {
     this.url = url;
   }
 
-  public ImageSample version(Integer version) {
+  public ImageSample version(Long version) {
     this.version = version;
     return this;
   }
@@ -80,11 +81,11 @@ public class ImageSample {
    * @return version
   **/
   @ApiModelProperty(value = "Auto-incrementing object version number.")
-  public Integer getVersion() {
+  public Long getVersion() {
     return version;
   }
 
-  public void setVersion(Integer version) {
+  public void setVersion(Long version) {
     this.version = version;
   }
 
@@ -131,6 +132,6 @@ public class ImageSample {
     }
     return o.toString().replace("\n", "\n    ");
   }
-  
+
 }
