+++ src/main/java/com/vendhq/models/InventoryCountItem.java
@@ -14,6 +14,7 @@
 package com.vendhq.models;
 
 import java.util.Objects;
+import java.util.Arrays;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
@@ -57,7 +58,7 @@ public class InventoryCountItem {
   private String deletedAt = null;
 
   @JsonProperty("version")
-  private Integer version = null;
+  private Long version = null;
 
   public InventoryCountItem productId(UUID productId) {
     this.productId = productId;
@@ -239,7 +240,7 @@ public class InventoryCountItem {
     this.deletedAt = deletedAt;
   }
 
-  public InventoryCountItem version(Integer version) {
+  public InventoryCountItem version(Long version) {
     this.version = version;
     return this;
   }
@@ -249,11 +250,11 @@ public class InventoryCountItem {
    * @return version
   **/
   @ApiModelProperty(required = true, value = "Auto-incrementing object version number.")
-  public Integer getVersion() {
+  public Long getVersion() {
     return version;
   }
 
-  public void setVersion(Integer version) {
+  public void setVersion(Long version) {
     this.version = version;
   }
 
@@ -316,6 +317,6 @@ public class InventoryCountItem {
     }
     return o.toString().replace("\n", "\n    ");
   }
-  
+
 }
