+++ src/main/java/com/vendhq/models/Outlet.java
@@ -14,6 +14,7 @@
 package com.vendhq.models;
 
 import java.util.Objects;
+import java.util.Arrays;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
@@ -71,7 +72,7 @@ public class Outlet {
   private String deletedAt = null;
 
   @JsonProperty("version")
-  private Integer version = null;
+  private Long version = null;
 
   public Outlet id(String id) {
     this.id = id;
@@ -343,7 +344,7 @@ public class Outlet {
     this.deletedAt = deletedAt;
   }
 
-  public Outlet version(Integer version) {
+  public Outlet version(Long version) {
     this.version = version;
     return this;
   }
@@ -353,11 +354,11 @@ public class Outlet {
    * @return version
   **/
   @ApiModelProperty(value = "Auto-incrementing object version number.")
-  public Integer getVersion() {
+  public Long getVersion() {
     return version;
   }
 
-  public void setVersion(Integer version) {
+  public void setVersion(Long version) {
     this.version = version;
   }
 
@@ -430,6 +431,6 @@ public class Outlet {
     }
     return o.toString().replace("\n", "\n    ");
   }
-  
+
 }
