+++ src/main/java/com/vendhq/models/PriceBook.java
@@ -14,6 +14,7 @@
 package com.vendhq.models;
 
 import java.util.Objects;
+import java.util.Arrays;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
@@ -54,7 +55,7 @@ public class PriceBook {
   private CustomerGroup customerGroup = null;
 
   @JsonProperty("version")
-  private Integer version = null;
+  private Long version = null;
 
   @JsonProperty("deleted_at")
   private String deletedAt = null;
@@ -221,7 +222,7 @@ public class PriceBook {
     this.customerGroup = customerGroup;
   }
 
-  public PriceBook version(Integer version) {
+  public PriceBook version(Long version) {
     this.version = version;
     return this;
   }
@@ -231,11 +232,11 @@ public class PriceBook {
    * @return version
   **/
   @ApiModelProperty(value = "Auto-incrementing object version number.")
-  public Integer getVersion() {
+  public Long getVersion() {
     return version;
   }
 
-  public void setVersion(Integer version) {
+  public void setVersion(Long version) {
     this.version = version;
   }
 
@@ -316,6 +317,6 @@ public class PriceBook {
     }
     return o.toString().replace("\n", "\n    ");
   }
-  
+
 }
