+++ src/main/java/com/vendhq/models/Version.java
@@ -14,6 +14,7 @@
 package com.vendhq.models;
 
 import java.util.Objects;
+import java.util.Arrays;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
@@ -27,12 +28,12 @@ import io.swagger.annotations.ApiModelProperty;
 
 public class Version {
   @JsonProperty("min")
-  private Integer min = null;
+  private Long min = null;
 
   @JsonProperty("max")
-  private Integer max = null;
+  private Long max = null;
 
-  public Version min(Integer min) {
+  public Version min(Long min) {
     this.min = min;
     return this;
   }
@@ -42,15 +43,15 @@ public class Version {
    * @return min
   **/
   @ApiModelProperty(required = true, value = "Lowest version number of the payload.")
-  public Integer getMin() {
+  public Long getMin() {
     return min;
   }
 
-  public void setMin(Integer min) {
+  public void setMin(Long min) {
     this.min = min;
   }
 
-  public Version max(Integer max) {
+  public Version max(Long max) {
     this.max = max;
     return this;
   }
@@ -60,11 +61,11 @@ public class Version {
    * @return max
   **/
   @ApiModelProperty(required = true, value = "Highest version number of the payload.")
-  public Integer getMax() {
+  public Long getMax() {
     return max;
   }
 
-  public void setMax(Integer max) {
+  public void setMax(Long max) {
     this.max = max;
   }
 
@@ -109,6 +110,6 @@ public class Version {
     }
     return o.toString().replace("\n", "\n    ");
   }
-  
+
 }
